dissolve PathPatternOptions and PathPatternOption, merge with PathPatternBuilder
